+++ tx/models/__init__.py
@@ -1,6 +1,6 @@
 from tx.models.mnist import Mnist
 from tx.models.qwen3 import Qwen3ForCausalLM
-from tx.models.qwen3_moe import Qwen3MoeForCausalLM
+from tx.models.qwen3 import Qwen3ForCausalLM as Qwen3MoeForCausalLM
 
 __all__ = [
     Mnist,
