+++ tx/models/qwen3.py
@@ -110,6 +110,76 @@ class Qwen3MLP(nnx.Module):
 
     def __call__(self, x: jax.Array) -> jax.Array:
         return self.down_proj(nnx.silu(self.gate_proj(x)) * self.up_proj(x))
+
+
+class Qwen3Experts(nnx.Module):
+
+    def __init__(self, config: Qwen3Config, *, dtype: jnp.dtype, rngs: nnx.Rngs) -> None:
+        self.config = config
+        self.gate_proj = Param(
+            config.num_experts, config.hidden_size, config.moe_intermediate_size,
+            dtype=dtype,
+            kernel_init=nnx.with_partitioning(nnx.initializers.lecun_normal(), jax.P(None, None, "tp")),
+            rngs=rngs
+        )
+        self.up_proj = Param(
+            config.num_experts, config.hidden_size, config.moe_intermediate_size,
+            dtype=dtype,
+            kernel_init=nnx.with_partitioning(nnx.initializers.lecun_normal(), jax.P(None, None, "tp")),
+            rngs=rngs
+        )
+        self.down_proj = Param(
+            config.num_experts, config.moe_intermediate_size, config.hidden_size,
+            dtype=dtype,
+            kernel_init=nnx.with_partitioning(nnx.initializers.lecun_normal(), jax.P(None, "tp", None)),
+            rngs=rngs
+        )
+
+    def __call__(self, hidden_states: jax.Array, router_logits: jax.Array) -> jax.Array:
+        # Get top-k experts for each token and compute routing weights
+        routing_weights, selected_experts = jax.lax.top_k(
+            router_logits, k=self.config.num_experts_per_tok
+        )
+        routing_weights = nnx.softmax(routing_weights, axis=-1)
+
+        # Prepare for ragged_dot by sorting tokens based on their assigned expert
+        selected_experts_flat = selected_experts.ravel()
+        hidden_states_expanded = jnp.repeat(hidden_states, self.config.num_experts_per_tok, axis=0)
+        sort_indices = jnp.argsort(selected_experts_flat)
+        hidden_states_sorted = hidden_states_expanded[sort_indices]
+        group_sizes = jnp.bincount(selected_experts_flat, length=self.config.num_experts)
+
+        # Apply expert layers using ragged_dot
+        gate_out = jax.lax.ragged_dot(hidden_states_sorted, self.gate_proj.value, group_sizes)
+        up_out = jax.lax.ragged_dot(hidden_states_sorted, self.up_proj.value, group_sizes)
+        down_out = jax.lax.ragged_dot(
+            nnx.silu(gate_out) * up_out, self.down_proj.value, group_sizes
+        )
+
+        # Unsort and combine the expert outputs
+        unsort_indices = jnp.argsort(sort_indices)
+        unsorted_out = down_out[unsort_indices]
+        reshaped_out = unsorted_out.reshape(-1, self.config.num_experts_per_tok, self.config.hidden_size)
+        return jnp.sum(reshaped_out * routing_weights[..., None], axis=1)
+
+
+class Qwen3MoeSparseMoeBlock(nnx.Module):
+
+    def __init__(self, config: Qwen3Config, *, dtype: jnp.dtype, rngs: nnx.Rngs) -> None:
+        self.config = config
+        self.gate = nnx.Linear(
+            config.hidden_size, config.num_experts,
+            use_bias=False, dtype=dtype, param_dtype=dtype,
+            kernel_init=nnx.with_partitioning(nnx.initializers.lecun_normal(), jax.P(None, "tp")), rngs=rngs,
+        )
+        self.experts = Qwen3Experts(config, dtype=dtype, rngs=rngs)
+
+    def __call__(self, hidden_states: jax.Array) -> tuple[jax.Array, jax.Array]:
+        original_shape = hidden_states.shape
+        hidden_states = hidden_states.reshape(-1, self.config.hidden_size)
+        router_logits = self.gate(hidden_states)
+        hidden_states = self.experts(hidden_states, router_logits)
+        return hidden_states.reshape(original_shape), router_logits
         
 
 class Qwen3DecoderLayer(nnx.Module):
